+++ library/Gcm/Message.php
@@ -34,9 +34,19 @@ class Message
     protected $collapseKey;
 
     /**
+     * @var string
+     */
+    protected $priority = 'normal';
+
+    /**
      * @var array
      */
     protected $data = [];
+
+    /**
+     * @var array
+     */
+    protected $notification = [];
 
     /**
      * @var bool
@@ -150,7 +160,33 @@ class Message
     }
 
     /**
-     * Set Data.
+     * Get priority
+     *
+     * @return string
+     */
+    public function getPriority()
+    {
+        return $this->priority;
+    }
+
+    /**
+     * Set priority
+     *
+     * @param string $priority
+     * @return Message
+     * @throws Exception\InvalidArgumentException
+     */
+    public function setPriority($priority)
+    {
+        if (!is_null($priority) && !(is_string($priority) && strlen($priority) > 0)) {
+            throw new Exception\InvalidArgumentException('$priority must be null or a non-empty string');
+        }
+        $this->priority = $priority;
+        return $this;
+    }
+
+    /**
+     * Set Data
      *
      * @param array $data
      *
@@ -215,7 +251,65 @@ class Message
     }
 
     /**
-     * Set Delay While Idle.
+     * Set notification
+     *
+     * @param array $data
+     * @return Message
+     */
+    public function setNotification(array $data)
+    {
+        $this->clearNotification();
+        foreach ($data as $k => $v) {
+            $this->addNotification($k, $v);
+        }
+        return $this;
+    }
+
+    /**
+     * Get notification
+     *
+     * @return array
+     */
+    public function getNotification()
+    {
+        return $this->notification;
+    }
+
+    /**
+     * Add notification data
+     *
+     * @param string $key
+     * @param mixed $value
+     * @return Message
+     * @throws Exception\InvalidArgumentException
+     * @throws Exception\RuntimeException
+     */
+    public function addNotification($key, $value)
+    {
+        if (!is_string($key) || empty($key)) {
+            throw new Exception\InvalidArgumentException('$key must be a non-empty string');
+        }
+        if (array_key_exists($key, $this->notification)) {
+            throw new Exception\RuntimeException('$key conflicts with current set data');
+        }
+        $this->notification[$key] = $value;
+        return $this;
+    }
+
+    /**
+     * Clear notification
+     *
+     * @return Message
+     */
+    public function clearNotification()
+    {
+        $this->notification = [];
+
+        return $this;
+    }
+
+    /**
+     * Set Delay While Idle
      *
      * @param bool $delay
      *
@@ -331,8 +425,14 @@ class Message
         if ($this->collapseKey) {
             $json['collapse_key'] = $this->collapseKey;
         }
+        if ($this->priority) {
+            $json['priority'] = $this->priority;
+        }
         if ($this->data) {
             $json['data'] = $this->data;
+        }
+        if ($this->notification) {
+            $json['notification'] = $this->notification;
         }
         if ($this->delayWhileIdle) {
             $json['delay_while_idle'] = $this->delayWhileIdle;