+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file, in reverse 
 
 - [#27](https://github.com/zendframework/ZendService_Google_Gcm/pull/27) PSR-4 schema
 - [#27](https://github.com/zendframework/ZendService_Google_Gcm/pull/27) PHP >= 5.5 & 7
+- [#20](https://github.com/zendframework/ZendService_Google_Gcm/pull/25) Notification and priority parameters for FCM
 
 ### Deprecated
 
@@ -22,6 +23,7 @@ All notable changes to this project will be documented in this file, in reverse 
 - [#27](https://github.com/zendframework/ZendService_Google_Gcm/pull/27) Fix travis CI integration
 - [#27](https://github.com/zendframework/ZendService_Google_Gcm/pull/27) Fix coding style (use ::class and short arrays)
 - [#27](https://github.com/zendframework/ZendService_Google_Gcm/pull/27) Fix docblocks for IDE integration
+- [#20](https://github.com/zendframework/ZendService_Google_Gcm/pull/25) Change endpoint to FCM
 
 ## 1.0.4 - TBD